+++ index.js
@@ -1,0 +1,7 @@
+import navigation from './modules/navigation.js';
+import library from './modules/library.js';
+import timeNow from './modules/date.js';
+
+navigation();
+library();
+timeNow();
